Move mongoose and elasticSearch to it own files, for better fine tuning in future.
Also create mongoose models on exact `db`. It allows to have several
mongodb connection for different models.
